+++ weaver/wps_restapi/jobs/utils.py
@@ -2,7 +2,7 @@ import math
 import os
 import shutil
 from copy import deepcopy
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
 
 import colander
 from celery.utils.log import get_task_logger
@@ -259,7 +259,10 @@ def get_schema_query(schema, strict=True):
     if not schema:
         return None
     # unescape query (eg: "OGC+strict" becomes "OGC string" from URL parsing)
-    schema_checked = str(schema).replace(" ", "+").lower()
+    schema_checked = cast(
+        "JobInputsOutputsSchemaType",
+        str(schema).replace(" ", "+").lower()
+    )
     if JobInputsOutputsSchema.get(schema_checked) is None:
         raise HTTPBadRequest(json={
             "type": "InvalidParameterValue",
@@ -352,57 +355,80 @@ def get_results(job,                                # type: Job
         # Literal result is only by itself. Therefore, find applicable field by non 'data' match.
         rtype = "href" if get_any_value(result, key=True, file=True, data=False) else "data"
         value = get_any_value(result)
-        out_key = rtype
-        out_id = get_any_id(result)
-        out_mode = out_ref.get(out_id, {}).get("transmissionMode")
-        as_ref = link_references and out_mode == ExecuteTransmissionMode.REFERENCE
-        if rtype == "href":
-            # fix paths relative to instance endpoint, but leave explicit links as is (eg: S3 bucket, remote HTTP, etc.)
-            if value.startswith("/"):
-                value = str(value).lstrip("/")
-            if "://" not in value:
-                value = wps_url + value
-        elif ogc_api:
-            out_key = "value"
-        elif value_key:
-            out_key = value_key
-        output = {out_key: value}
-        if rtype == "href":  # required for the rest to be there, other fields optional
-            if "mimeType" not in result:
-                result["mimeType"] = get_format(value, default=ContentType.TEXT_PLAIN).mime_type
-            if ogc_api or not strict:
-                output["type"] = result["mimeType"]
-            if not ogc_api or not strict or as_ref:
-                output["format"] = {fmt_key: result["mimeType"]}
-                for field in ["encoding", "schema"]:
-                    if field in result:
-                        output["format"][field] = result[field]
-        elif rtype != "href":
-            dtype = result.get("dataType", any2wps_literal_datatype(value, is_value=True) or "string")
-            if ogc_api:
-                output["dataType"] = {"name": dtype}
-            else:
-                output["dataType"] = dtype
-
-        if schema == JobInputsOutputsSchema.OGC_STRICT:
-            out_fmt = output.pop("format", {})
-            for fmt_key, fmt_val in out_fmt.items():
-                output.setdefault(fmt_key, fmt_val)
-
-        if ogc_api or as_ref:
-            mapping = references if as_ref else outputs
-            if out_id in mapping:
-                output_list = mapping[out_id]
-                if not isinstance(output_list, list):
-                    output_list = [output_list]
-                output_list.append(output)
-                mapping[out_id] = output_list
-            else:
-                mapping[out_id] = output
+        # An array of literals can be merged as-is
+        if (
+            isinstance(value, list) and
+            all(
+                isinstance(val, (bool, float, int, str, type(None)))
+                for val in value
+            )
+        ):
+            array = [value]  # array of array such that it iterated as the array of literals directly
+        # Any other type of array implies complex data (bbox or file)
+        # They must be defined on their own with respective media-type/format details per item.
         else:
-            # if ordered insert supported by python version, insert ID first
-            output = dict([("id", out_id)] + list(output.items()))  # noqa
-            outputs.append(output)
+            array = value if isinstance(value, list) else [value]
+        for val_item in array:
+            val_data = val_item
+            if isinstance(val_item, dict) and isinstance(value, list):
+                rtype = "href" if get_any_value(val_item, key=True, file=True, data=False) else "data"
+                val_data = get_any_value(val_item, file=True, data=False)
+            out_key = rtype
+            out_id = get_any_id(result)
+            out_mode = out_ref.get(out_id, {}).get("transmissionMode")
+            as_ref = link_references and out_mode == ExecuteTransmissionMode.REFERENCE
+            if rtype == "href" and isinstance(val_data, str):
+                # fix paths relative to instance endpoint,
+                # but leave explicit links as is (eg: S3 bucket, remote HTTP, etc.)
+                if val_data.startswith("/"):
+                    val_data = val_data.lstrip("/")
+                if "://" not in val_data:
+                    val_data = wps_url + val_data
+            elif ogc_api:
+                out_key = "value"
+            elif value_key:
+                out_key = value_key
+
+            output = {out_key: val_data}
+
+            # required for the rest to be there, other fields optional
+            if rtype == "href":
+                val_fmt = val_item if isinstance(val_item, dict) else result
+                if "mimeType" not in val_fmt:
+                    val_fmt["mimeType"] = get_format(val_data, default=ContentType.TEXT_PLAIN).mime_type
+                if ogc_api or not strict:
+                    output["type"] = val_fmt["mimeType"]
+                if not ogc_api or not strict or as_ref:
+                    output["format"] = {fmt_key: val_fmt["mimeType"]}
+                    for field in ["encoding", "schema"]:
+                        if field in result:
+                            output["format"][field] = val_fmt[field]
+            elif rtype != "href":
+                dtype = result.get("dataType", any2wps_literal_datatype(val_data, is_value=True) or "string")
+                if ogc_api:
+                    output["dataType"] = {"name": dtype}
+                else:
+                    output["dataType"] = dtype
+
+            if schema == JobInputsOutputsSchema.OGC_STRICT:
+                out_fmt = output.pop("format", {})
+                for fmt_key, fmt_val in out_fmt.items():
+                    output.setdefault(fmt_key, fmt_val)
+
+            if ogc_api or as_ref:
+                mapping = references if as_ref else outputs
+                if out_id in mapping:
+                    output_list = mapping[out_id]
+                    if not isinstance(output_list, list):
+                        output_list = [output_list]
+                    output_list.append(output)
+                    mapping[out_id] = output_list
+                else:
+                    mapping[out_id] = output
+            else:
+                # if ordered insert supported by python version, insert ID first
+                output = dict([("id", out_id)] + list(output.items()))  # noqa
+                outputs.append(output)
 
     # needed to collect and aggregate outputs of same ID first in case of array
     # convert any requested link references using indices if needed